api/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
rich==13.8.1
supabase==2.7.4
stripe==10.12.0
python-dotenv==1.0.1